test_run_analyze_trades.py
import run_analyze_trades as analyze_trades
import pytest
import mock
import sys
from xml.dom import minidom
import pandas as pd


@pytest.fixture()
def source_xml_file(tmpdir):
    """ Fixture creates a XML source file with some data.

    This should mimin the data expected to be found in the actual XML source.
    """
    xml_input = tmpdir.mkdir('sub').join('trades_raw.xml')
    xml_input.write(
        '''
        <Trades>
            <Trade CorrelationId="701" NumberOfTrades="1" Limit="1000" TradeID="A1">700</Trade>
            <Trade CorrelationId="002" NumberOfTrades="1" Limit="1000" TradeID="B2">1170</Trade>
            <Trade CorrelationId="103" NumberOfTrades="2" Limit="500"  TradeID="C3">200</Trade>
        </Trades>
        '''
    )
    return str(xml_input)


@pytest.fixture()
def dataframe_raw_trades():
    """ Returns a dataframe of raw trades based on the source XML fixture.
    """
    data = [
        ['701', '1', '1000', '700',  'A1'],
        ['002', '1', '1000', '1170', 'B2'],
        ['103', '2', '500',  '200',  'C3']
    ]
    df_cols = ['CorrelationID', 'NumberOfTrades', 'Limit', 'Value', 'TradeID']
    df = pd.DataFrame(data, columns=df_cols)
    return df


@pytest.fixture()
def dataframe_aggregated_trades():
    """ Returns a dataframe of aggregated trade values per correlationID,
    and sorted by correlationID
    """
    data = [
        ['002', '1', '1000', '1170', 'Rejected'],
        ['103', '2', '500',  '200',  'Pending'],
        ['701', '1', '1000', '700',  'Accepted'],
    ]
    df_cols = ['CorrelationID', 'NumberOfTrades', 'Limit', 'Value', 'State']
    df = pd.DataFrame(data, columns=df_cols)
    df[['NumberOfTrades', 'Limit', 'Value']] = df[
        ['NumberOfTrades', 'Limit', 'Value']].apply(pd.to_numeric)
    return df


class TestRunAnalyzeTrades(object):

    def test_read_xml_input_file(self, source_xml_file):
        sys.argv = ['foo.py', source_xml_file]
        result = analyze_trades.read_xml_input_file()
        assert isinstance(result, minidom.Document)

    def test_parse_trades_from_xml_data(self, source_xml_file):
        doc = minidom.parse(source_xml_file)
        result = analyze_trades.parse_trades_from_xml_data(doc)
        assert isinstance(result, list)
        assert len(result) == 3

    def test_get_dataframe_of_all_input_trades(self, source_xml_file,  dataframe_raw_trades):
        doc = minidom.parse(source_xml_file)
        dom_trades = analyze_trades.parse_trades_from_xml_data(doc)
        result = analyze_trades.get_dataframe_of_all_input_trades(dom_trades)
        pd.util.testing.assert_frame_equal(result, dataframe_raw_trades)

    @pytest.mark.xfail(reason='dtypes differing in assertion', strict=True)
    def test_get_dataframe_of_aggregated_trades(self, dataframe_raw_trades, dataframe_aggregated_trades):
        result = analyze_trades.get_dataframe_of_aggregated_trades(dataframe_raw_trades)
        pd.util.testing.assert_frame_equal(result, dataframe_aggregated_trades)

    @pytest.mark.skip()
    def test_output_accepted_rejected_trades_to_csv(self):
        # TODO ran out of time
        pass

    def test_check_cmdline_args_for_valid_args(self, ):
        sys.argv = ['foo.py', 'bar.xml']
        result = analyze_trades._check_cmdline_args()
        assert result == 'bar.xml'

    def test_check_cmdline_args_for_invalid_args(self):
        sys.argv = ['foo.py']
        with pytest.raises(ValueError):
            analyze_trades._check_cmdline_args()
